First commit
This first commit is a rather large one. I forgot to commit as I went today. I've got to get better at doing that. This includes the initialization of rspec and the creation of the lib folder. Then I wrote the tests for the player, piece, and board classes, and wrote the code for those classes. The board class turned out to be pretty difficult to write and test for because of the how the win conditions of the game work. I first set up the board as a hash with each key representing a position (eg A1, A2, C5, etc.). But in the end I changed the board to be an array of arrays with each inner array representing a row on the board.