Finalize Game tests and code. Update board tests and code. Create board_design.txt. Create next_steps.txt
This is my final commit for now on this project. Its mostly complete. The only remaining steps are to finish the board design and clean up the Board#valid_column? method (details are in next_steps.txt). I'm choosing to finish the project at this point because the goal of this project is to learn Rspec, which I feel I've done a pretty good job of, and the remaining work doesn't require much in terms of testing.